Update profit report

+++ b1ing-service/src/main/java/com/game/b1ingservice/postgres/jdbc/dto/ProfitReport.java
@@ -2,8 +2,10 @@ package com.game.b1ingservice.postgres.jdbc.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProfitReport {
     private Integer labels;
-    private Integer data;
+    private BigDecimal data;
 }
